MERGE: `gui` into `dev`

+++ gui/include/graphical/Graphics.hpp
@@ -10,23 +10,24 @@
     #define GRAPHICS_HPP
     #include <cstddef>
     #include <memory>
-#include <queue>
-#include <string>
+    #include <queue>
+    #include <string>
     #include <unordered_map>
-#include "AssetsManager.hpp"
+    #include "AssetsManager.hpp"
     #include "GameInfo.hpp"
-#include "MenuInfo.hpp"
+    #include "MenuInfo.hpp"
     #include "raylib.h"
     #include <cstring>
-#include <vector>
-#include "QueueManager.hpp"
+    #include <vector>
+    #include "QueueManager.hpp"
+    #include "VarManager.hpp"
 
 
 
 namespace Gui {
     constexpr size_t WINDOW_WIDTH = 1200;
     constexpr size_t WINDOW_HEIGHT = 600;
-    constexpr size_t TILE_SIZE = 5;
+    constexpr float TILE_SIZE = 5.0;
 
     class Graphics {
         public:
@@ -41,7 +42,6 @@ namespace Gui {
                 SCOREBOARD,
             };
         public:
-            // Modifiez le constructeur pour prendre QueueManager au lieu des queues
             Graphics(std::shared_ptr<QueueManager> queueManager);
             ~Graphics();
 
@@ -58,6 +58,7 @@ namespace Gui {
         ////////////////////////////////////// GRAPHIC //////////////////////////////////////
             void init(void);
             void run(std::atomic<bool> &);
+
         private:
             void handleEvents(void);
             void update(void);
@@ -97,7 +98,6 @@ namespace Gui {
             std::shared_ptr<MenuInfo> _menu;
             std::shared_ptr<GameInfo> _game;
             AssetsManager _assetsManager;
-            // int = temporaire
     };
 }
